replace fonts with Terminus, display all Z80 registers, misc stuff
Replace the fonts with a Sans-Serif one. I choose Terminus.

Since this font is not as wide as the old Courier New, we now
have space to display all Z80 registers.

Declare refresh_flag and cpudisp_flag as "volatile", since these
variables are accessed by an asynchronously running function.

Increase SPI frequency to 40MHz. That is well within the specs of the
LCD controller, according to the datasheet.

Add 8-bit support to the GUI library, not that we could use it, but
I had an itch, and it was trivial to implement.

Optimize Paint_Clear a bit more.

Mark Paint_SetPixel and Paint_DrawChar as __time_critical_func, so
that they execute from RAM.

We can now run at 30Hz LCD refresh frequency.