flesh out fonts to 128 chars, some drawing code tweaking
Flesh out fonts to 128 characters, so that the drawing code doesn't
draw some random memory bits if someone passes it non-printable ones.

Add Font22 (overlooked in the initial commit of Terminus fonts).

Add more metrics to the sFONT type, so that they don't need
to be calculated.

Some performance tweaks to the drawing code.