use DMA transfer with interrupt for the LCD refresh
Now runs with 60Hz even on the RP2040. The RP2350 can also use
60Hz with LCD_COLOR_DEPTH set to 16.

Add DMA_IRQ_num initializer to sdio_if even if it is the default,
so we know it uses DMA_IRQ_0. The LCD code uses DMA_IRQ_1.

Moved the LCD initialization/deinitialization code into lcd_task(),
since the interrupt setup must be run on the core that will receive
the interrupts.