redo RGB LED, fix load file command 'r' to set the PC to 0
Disk read blinks the green component, write the red component.

Added output port 0 to turn the blue component of the LED on/off.

Added blink.asm to src-examples.

Changed load_file() to return 0/1 on error/success.

Set the PC to 0 on successful load of a file with 'r'.

Use the same info line for CPU registers and LED panel displays.